Update Depth3D.fx
I think I came up with a "safer way," of allowing the debug view of the depth buffer.

So many issue with loops....
DX12 nor Vulkan was tested.
Do-While Loop Seems to be faster then for or while loop in DX 9, 10, and 11. But, not in openGL. In some rare openGL games it causes CTD
For loop is broken in this shader for some reason in DX9. I don't know why. This is the reason for the change.
While Loop is the most compatible of the bunch. So I am forced to use this loop.

DX9 why can't you be just be normal?