Adding one shader to the Nvidia Shader Pool.
Flair.fx is a "Lens Flare" based shader that can use depth to assist in the application of the flares.
Not all lens flare system is emulated by this shader. Some other flares can and may be added later.
This shader was made for Nvidia Freestyle system and can also be used in normal ReShade. 

I am also uploading this shader because I wanted to keep this archive a bit fresh. 

Thank you, Handpunk for support.
Thank you, crosire for making this possible and are you sure we don't capitalize your name? :smile: