More user friendly UI. Separate kernel widths for X and Y axis.
Changed the variable names and gaussian kernel width calculation to be more user friendly. i.e. a width of 0.5 leading to a gaussian kernel with standard deviation of 0.5. And number of taps from 7 to 35 affecting accuracy only, not radius.

Added separate kernel widths for X and Y axis. i.e. interlaced content can benefit from a higher width in the Y axis.